requirements_weaviate.txt
weaviate-client>=4.3.1
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
openai>=1.0.0
Pillow>=9.0.0
requests>=2.28.0
transformers>=4.20.0
torch>=1.12.0
cohere>=4.5.0